+++ .gitignore
@@ -39,6 +39,6 @@ out/
 ### core:core-notification ###
 core/core-notification/src/main/resources/firebase/*.json
 
-### support:jwt.crypto ###
+### core:core-auth.support.crypto ###
 *.pem
-support/jwt/src/main/resources/*.properties
+core/core-auth/src/main/resources/*.properties
