+++ core/core-auth/build.gradle
@@ -1,3 +1,5 @@
 dependencies {
     compileOnly("org.springframework:spring-context")
+    implementation 'com.auth0:java-jwt:4.4.0'
+
 }
