--- core/core-auth/src/main/java/com/barlow/core/auth/authentication/access/MemberAuthenticationService.java
@@ -1,18 +1,0 @@
-package com.barlow.core.auth.authentication.access;
-
-import com.barlow.core.auth.authentication.token.Token;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MemberAuthenticationService {
-
-    private final MemberAccessTokenAuthenticator memberTokenAuthenticator;
-
-    public MemberAuthenticationService(MemberAccessTokenAuthenticator memberTokenAuthenticator) {
-        this.memberTokenAuthenticator = memberTokenAuthenticator;
-    }
-
-    public MemberPrincipal authenticate(Token token) {
-        return memberTokenAuthenticator.authenticate(token);
-    }
-}
