+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/token/MemberAccessPayload.java
@@ -4,7 +4,4 @@ public record MemberAccessPayload(
         Long memberNo,
         String role
 ) {
-    public static MemberAccessPayload invalid() {
-        return new MemberAccessPayload(-1L, "");
-    }
 }
