--- core/core-auth/src/main/java/com/barlow/core/auth/authentication/token/MemberTokenValidator.java
@@ -1,9 +1,0 @@
-package com.barlow.core.auth.authentication.token;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public interface MemberTokenValidator {
-
-    TokenValidationResult<MemberAccessPayload> validate(Token token);
-}
