--- core/core-auth/src/main/java/com/barlow/core/auth/authentication/token/TokenValidationResultType.java
@@ -1,8 +1,0 @@
-package com.barlow.core.auth.authentication.token;
-
-public enum TokenValidationResultType {
-
-    FAILURE,
-    EXPIRED,
-    SUCCESS
-}
