--- support/jwt/build.gradle
@@ -1,7 +1,0 @@
-dependencies {
-    implementation project(':core:core-auth')
-    implementation 'com.auth0:java-jwt:4.4.0'
-    implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
-
-    compileOnly("org.springframework:spring-context")
-}
