--- support/jwt/src/test/java/com/barlow/support/jwt/token/JwtTest.java
@@ -1,9 +1,0 @@
-package com.barlow.support.jwt.token;
-
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-
-@SpringBootTest
-@Import(TestKeyConfig.class)
-public abstract class JwtTest {
-}
